Resolve some issues
