+++ .idea/gradle.xml
@@ -15,6 +15,5 @@
         </option>
       </GradleProjectSettings>
     </option>
-    <option name="offlineMode" value="true" />
   </component>
 </project>
