+++ app/.gitignore
@@ -1,1 +1,10 @@
+*.iml
+.gradle
+.idea
+/local.properties
+.DS_Store
 /build
+/captures
+.externalNativeBuild
+.cxx
+local.properties
