+++ app/build.gradle
@@ -2,6 +2,7 @@ plugins {
     id 'com.android.application'
     id 'kotlin-android'
     id 'kotlin-kapt'
+    id 'dagger.hilt.android.plugin'
 }
 
 android {
@@ -64,6 +65,9 @@ dependencies {
 
     implementation 'ch.acra:acra-http:5.9.5'
     implementation 'ch.acra:acra-dialog:5.9.5'
+
+    implementation "com.google.dagger:hilt-android:2.38.1"
+    kapt "com.google.dagger:hilt-compiler:2.38.1"
 
     implementation 'com.google.code.gson:gson:2.8.9'
     implementation 'com.github.ark-builders:ark-filepicker:main-SNAPSHOT'
