--- app/src/main/kotlin/com/ark/globe/coordinates/Locations.kt
@@ -1,3 +1,0 @@
-package com.ark.globe.coordinates
-
-data class Locations(var locations : List<Location>)
